mise en place partielle des themes et gestion de l'historique en javascript

+++ app/models/level.rb
@@ -81,31 +81,33 @@ class Level < ActiveRecord::Base
   
   # generate an image of the level in /public/images/levels
   def generate_thumb
-    @pack  = self.pack
-    @level = self
-    @grid  = self.grid_with_floor
-    
-    @grid.each_with_index do |row, index|
-      row = row.gsub(' ', 'empty64:png ')
-      row = row.gsub('s', 'floor64:png ')
-      row = row.gsub('.', 'goal64:png ')
-      row = row.gsub('#', 'wall64:png ')
-      row = row.gsub('$', 'box64:png ')
-      row = row.gsub('*', 'boxgoal64:png ')
-      row = row.gsub('@', 'pusher64:png ')
-      row = row.gsub('+', 'pushergoal64:png ')
-      row = row.gsub(':', '.')
+    unless File.exist?("public/images/levels/#{self.id}.png")
+      @pack  = self.pack
+      @level = self
+      @grid  = self.grid_with_floor
       
-      `cd public/images/;convert #{row} +append levels/row_#{index+1}.png`
+      @grid.each_with_index do |row, index|
+        row = row.gsub(' ', 'empty64:png ')
+        row = row.gsub('s', 'floor64:png ')
+        row = row.gsub('.', 'goal64:png ')
+        row = row.gsub('#', 'wall64:png ')
+        row = row.gsub('$', 'box64:png ')
+        row = row.gsub('*', 'boxgoal64:png ')
+        row = row.gsub('@', 'pusher64:png ')
+        row = row.gsub('+', 'pushergoal64:png ')
+        row = row.gsub(':', '.')
+              
+        `cd public/images/themes/classic;convert #{row} +append ../../levels/row_#{index+1}.png`
+      end
+      
+      command = ""
+      (1..self.height).to_a.each do |m|
+        command = command + "row_#{m}.png "
+      end
+      
+      `cd public/images/levels;convert #{command} -append #{self.id}.png`
+      `cd public/images/levels;rm -f row_*.png`
     end
-  
-    command = ""
-    (1..self.height).to_a.each do |m|
-      command = command + "row_#{m}.png "
-    end
-    
-    `cd public/images/levels;convert #{command} -append #{self.id}.png`
-    `cd public/images/levels;rm -f row_*.png`
   end
   
   private
